backend/requirements.txt
numpy
tqdm
langchain-google-genai 
python-dotenv
pandas
langchain-openai
yfinance
psycopg2-binary
langchain_tavily
matplotlib
seaborn
tabulate
emoji
langgraph
loguru
sqlglot
fastapi
uvicorn
adjustText